test: move TestPayloadServer
